Revert "wip extract getItemEditURL to utils"
This reverts commit ebb35f6001c374369c185a18432ede37d6962fd7.
