+++ packages/docusaurus-plugin-content-pages/src/plugin-content-pages.d.ts
@@ -8,11 +8,7 @@
 declare module '@docusaurus/plugin-content-pages' {
   import type {MDXOptions} from '@docusaurus/mdx-loader';
   import type {LoadContext, Plugin} from '@docusaurus/types';
-  import type {
-    FrontMatterLastUpdate,
-    LastUpdateData,
-    EditUrlFunction,
-  } from '@docusaurus/utils';
+  import type {FrontMatterLastUpdate, LastUpdateData} from '@docusaurus/utils';
 
   export type Assets = {
     image?: string;
@@ -63,6 +59,20 @@ declare module '@docusaurus/plugin-content-pages' {
     description?: string;
     unlisted: boolean;
   };
+
+  export type EditUrlFunction = (editUrlParams: {
+    /**
+     * The root content directory containing this post file, relative to the
+     * site path. Usually the same as `options.path` but can be localized
+     */
+    pagesDirPath: string;
+    /** Path to this pages file, relative to `pagesDirPath`. */
+    pagesPath: string;
+    /** @see {@link PagesPostMetadata.permalink} */
+    permalink: string;
+    /** Locale name. */
+    locale: string;
+  }) => string | undefined;
 
   export type Metadata = JSXPageMetadata | MDXPageMetadata;
